requirements.txt
# Ira - Requirements
# Generated from pyproject.toml dependencies
# Install with: pip install -r requirements.txt

# Core AI/ML
openai>=1.0.0
anthropic>=0.18.0

# Vector database
qdrant-client>=1.7.0

# Embeddings
voyageai>=0.2.0

# Web framework
requests>=2.31.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0

# PDF processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# Database
psycopg2-binary>=2.9.0

# Google APIs
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0

# Async
aiohttp>=3.9.0

# Environment
python-dotenv>=1.0.0

# Reranking
flashrank>=0.2.0

# Memory
mem0ai>=0.0.20

# Web framework for health endpoints
fastapi>=0.104.0
uvicorn>=0.24.0

# Telegram Bot
python-telegram-bot>=20.0

# PDF Generation (for quote automation)
fpdf2>=2.7.0

# Scheduling (for proactive outreach)
schedule>=1.2.0

# Templates (for dashboard)
jinja2>=3.1.0

# =============================================================================
# NEW DEPENDENCIES - IRA System Elevation Roadmap
# =============================================================================

# P0: Guardrails & PDF Extraction
# --------------------------------
# Camelot for high-accuracy table extraction from PDFs
# Requires: pip install camelot-py[cv] (includes opencv dependencies)
camelot-py[cv]>=0.11.0

# P1: Observability & Structured Logging
# --------------------------------------
# Structlog for production-grade structured logging
structlog>=24.0.0
# OpenTelemetry for distributed tracing
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
# OpenLLMetry for LLM observability (optional, uncomment if using)
# openllmetry>=0.0.30

# P2: Advanced RAG & NLU
# ----------------------
# RAGatouille for ColBERT-based reranking (more accurate than FlashRank)
# Set RERANKER_TYPE=colbert to enable
ragatouille>=0.0.8

# spaCy for fast NLU (intent classification, entity extraction)
spacy>=3.7.0
# Note: After installing, download the model:
# python -m spacy download en_core_web_sm

# Tenacity for production-grade retry logic
tenacity>=8.2.0

# pyrate-limiter for rate limiting external API calls
pyrate-limiter>=3.0.0

# python-docx for DOCX file extraction in knowledge engine
python-docx>=1.1.0

# Langfuse for LLM observability dashboard
# Set LANGFUSE_PUBLIC_KEY and LANGFUSE_SECRET_KEY to enable
langfuse>=2.0.0

# Redis for distributed caching (optional - falls back to in-memory)
redis>=5.0.0

# Async PostgreSQL client for health checks
asyncpg>=0.29.0
neo4j>=5.15.0
